% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract.R
\name{dir_to_files}
\alias{dir_to_files}
\title{Get all files in all sub-directories}
\usage{
dir_to_files(dir_path, regexp = ".*")
}
\arguments{
\item{regexp}{A regular expression (e.g. \verb{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}
}
\description{
Get all files in all sub-directories
}
\keyword{internal}
