% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_responses.R
\name{ensure_columns}
\alias{ensure_columns}
\title{Ensure that a data frame has the required columns or signal with a message}
\usage{
ensure_columns(data, required_columns, signal, message_prefix = "")
}
\arguments{
\item{data}{The \code{\link{data.frame}} to check.}

\item{required_columns}{A character vector of columns to ensure.}

\item{signal}{The method used to signal if any of the columns are missing (e.g. \code{\link{abort}}).}

\item{message_prefix}{A string to prepend to the message.}
}
\value{
A logical indicating if all the columns were present (\code{TRUE}) or not (\code{FALSE}).
}
\description{
Ensure that a data frame has the required columns or signal with a message
}
\keyword{internal}
