% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_responses.R
\name{ensure_data_in_responses}
\alias{ensure_data_in_responses}
\title{Ensure a responses data frame has the required columns.}
\usage{
ensure_data_in_responses(responses)
}
\arguments{
\item{responses}{A CourseKata responses table (e.g. from responses.csv).}
}
\value{
If no problems are found, the original responses table is returned as
a \code{\link{tibble}}. When problems are found, the corresponding rows are
removed from the output table.
}
\description{
A responses table isn't very useful without some basic information: the class
the response was elicited in, the student it was elicited from, and the
prompt that elicited the response (but not necessarily the response -- a
missing response is a response in itself). \code{ensure_data_in_responses()}
ensures that the columns exist in the data frame (failing with an error
message if not) and filters out rows that are missing required data (emitting
a warning if so).
}
\seealso{
process_responses
}
