% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_to_temp}
\alias{extract_to_temp}
\title{Extract a zip file to a temporary directory.}
\usage{
extract_to_temp(path, regexp = ".*")
}
\arguments{
\item{regexp}{Regular expressions for the files to extract from the zip file.}

\item{zip_file}{A character vector of the zip files to extract.}
}
\value{
A character vector of the directory where each zip file was extracted to.
}
\description{
Vectorized over path and pattern.
}
\keyword{internal}
