% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_zip_file}
\alias{is_zip_file}
\title{Does the path point to a zip file?}
\usage{
is_zip_file(path)
}
\arguments{
\item{path}{A character vector of paths to check.}
}
\value{
A logical vector indicating the paths that are zip files.
}
\description{
Does the path point to a zip file?
}
\keyword{internal}
