% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load file(s) from a directory, zipped directory, or file path.}
\usage{
load_data(object, regexp = ".*", class_id = NULL)
}
\arguments{
\item{object}{The object to process the data from. This can be one of a variety of options:
\itemize{
\item the path to a CourseKata data download zip file
\item the path to an extracted data download directory
\item the path to a specific CSV file (e.g. \code{classes.csv})
\item the \code{data.frame} loaded from a CSV file (useful for writing unit tests)
}}

\item{regexp}{If \code{object} is a directory or zipped directory, a \code{\link{regex}} pattern indicating
which file names to search for. The default value \code{'.*'} matches all files.}

\item{class_id}{(Optional) A character vector of the IDs of specific classes you want to extract
data for. These values should correspond to the names of the directories where the desired CSV
files can be found in the original data download.}
}
\value{
The resulting data frame will always be returned as a \code{\link{tibble}} with all factor
variables converted to character. If the \code{object} is a \code{data.frame} it will still be converted
to a \code{tibble}. If \code{object} is a path of some type, the files at the given location are filtered
against \code{regexp} and \code{class_id}, and then read in. As the files are read, the rows of each
table are merged int a single \code{tibble}.
}
\description{
Load file(s) from a directory, zipped directory, or file path.
}
