% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_responses.R
\name{map_response}
\alias{map_response}
\title{Map a response array to its values in a look-up table.}
\usage{
map_response(response, reference, lookup_table)
}
\arguments{
\item{response}{A text array of response options, e.g. "['1', '4']"}

\item{reference}{A reference that corresponds to a \code{lrn_question_reference}.}

\item{lookup_table}{A lookup table with a column \code{lrn_question_reference}, and a column for each
possible response option.}
}
\value{
The value(s) indicated by the response options, separated by semicolons if applicable.
}
\description{
If you are using this function you very likely understand why you need to use it and can figure
out what it does. It it is more likely that you want the more comprehensive function
\code{\link{process_responses}}. That said, give it a response, a question reference, and a look-up table,
and it will give you back the value from the look-up table that corresponds to the response.
}
\seealso{
process_responses
}
