% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Check if x is not in the look-up table}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\value{
A logical vector, indicating if a match was NOT located for each element of x: thus the
values are TRUE or FALSE and never NA.
}
\description{
Check if x is not in the look-up table
}
\seealso{
base::match
}
