% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parsers.R
\name{strip_alpha}
\alias{strip_alpha}
\alias{parse_double}
\alias{parse_integer}
\title{Convert character vectors to numeric vectors.}
\usage{
strip_alpha(x)

parse_double(x)

parse_integer(x)
}
\arguments{
\item{x}{The character vector to convert.}
}
\value{
For \code{parse_double} and \code{parse_integer} a vector only containing doubles or integers. The
helper function \code{strip_alpha} removes all non-numeric values leaving only numbers, dots (\code{.}
for decimals), and hyphens (\code{-} for negative values). The parsers will fail if \code{strip_alpha}
yields something non-numeric looking, like "1.1.1".
}
\description{
Given a character vector, these function will strip non-numeric characters from each item and
then attempt to convert the result to either a double using \code{\link[=as.double]{as.double()}} or an integer using
\code{\link[=as.integer]{as.integer()}}.
}
\keyword{internal}
