% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parsers.R
\name{parse_datetime}
\alias{parse_datetime}
\title{Parse a vector of date-time strings and convert to the desired time-zone.}
\usage{
parse_datetime(x, time_zone = "UTC")
}
\arguments{
\item{x}{The character vector to convert.}

\item{time_zone}{The time zone to use when parsing date-time objects (see \code{\link{timezones}}).}
}
\value{
A vector of date-times, where the date-time has been converted to the desired time-zone.
}
\description{
This function assumes that it is being given date-time strings in the UTC/GMT time-zone with a
specific format. The format is similar to ISO-8601, but it omits the "T" dividing the date and
time: \verb{YYYY-MM-DD hh:mm:ss.sss}. For example, the date-time May 1, 1983 at 3:33PM and 44.444
seconds, would be represented as \verb{1983-05-01 15:33.444}. After parsing those strings as
\link[=DateTimeClasses]{date-times}, the time-zone is converted to the desired zone.
}
\keyword{internal}
