% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_classes.R
\name{process_classes}
\alias{process_classes}
\title{Process information about classes in a CourseKata class data download}
\usage{
process_classes(path)
}
\arguments{
\item{path}{The path to a data download zip file, the path to an extracted
data download zip directory, or the path to a \code{classes.csv} file from
a data download.}
}
\value{
A \code{\link{tibble}} of the classes found in the given path. The
information is parsed from the first classes.csv file that is found (only
one should exist).
}
\description{
Process information about classes in a CourseKata class data download
}
