% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_responses.R
\name{process_responses}
\alias{process_responses}
\alias{convert_types_in_responses}
\alias{ensure_data_in_responses}
\alias{map_response_options}
\title{Process responses from a CourseKata class data download}
\usage{
process_responses(object, time_zone = "UTC", class_id = "")

convert_types_in_responses(responses, time_zone = "UTC")

ensure_data_in_responses(responses)

map_response_options(responses)
}
\arguments{
\item{object}{The path to a data download zip file, the path to an extracted
data download zip directory, the path to a \code{responses.csv} file from a
data download, or a responses-like \code{data.frame}.}

\item{time_zone}{The time zone to use when parsing date-times. See
\code{\link{timezones}} for more information about valid time zones.}

\item{class_id}{(Optional) A character vector of the IDs of specific classes
for which to extract responses. These values should correspond to the names
of the directories where the desired \code{responses.csv} files can be
found (this is how the zip file is structured when downloading from
CourseKata).}

\item{responses}{A CourseKata responses table (e.g. from responses.csv).}
}
\value{
A \code{\link{tibble}} of all responses found in the given path or
data frame. If there are multiple \code{responses.csv} files, they are
merged together and can be distinguished by the \code{class_id} column. If
\code{object} is a directory and \code{class_id} is specified, only the
responses for the specified class or classes are included.

This function returns the same responses table where the variables in
the table have all been converted to appropriate types. See the \strong{Details}
section for information about which variables are converted to which type.

If no problems are found, the original responses table is returned.
Otherwise, a responses table where all rows contain the required
informaiton.

The responses table with the lookup table added as the
\code{option_value_table} attribute and the \code{response} column altered
by mapping the responses using the table.
}
\description{
This function processes the responses found in the given object. The
responses are checked for required columns
(\code{\link{ensure_data_in_responses}}), the variables in the outcome table
are converted to appropriate types
(\code{\link{convert_types_in_responses}}), and multiple-response option
indices in the \code{response} column for \code{lrn_type = "mcq"} are mapped
to their values (in the \code{lrn_option_<index>} columns
(\code{\link{map_response_options}}).

A responses table isn't very useful without some basic information: the class
the response was elicited in, the student it was elicited from, and the
prompt that elicited the response (but not necessarily the response -- a
missing response is a response in itself). This function ensures that the
columns exist in the data frame (failing with an error message if not) and
filters out rows that are missing required data (emitting a warning if so).

Extract a lookup table for the multiple-choice questions in the set and then
use it to map response options to the values. The table is comprised of the
\code{lrn_question_reference} and \code{lrn_option_<option numbers>} columns,
and it is added to the resulting respnoses table as the
\code{option_value_table} attribute.
}
\details{
\describe{
\item{integer}{\code{attempt, lrn_question_position}}
\item{number}{\code{points_possible, points_earned}}
\item{datetime}{\code{dt_submitted, lrn_dt_started, lrn_dt_saved}}
\item{list}{\code{lrn_response_json}}
\item{character}{all other columns in the data frame not listed above}
}
}
