% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parsers.R
\name{safe_convert_json}
\alias{safe_convert_json}
\title{Convert a JSON-encoded vector to a list, ignoring all errors.}
\usage{
safe_convert_json(json_col)
}
\value{
A vector of lists, where each list has keys corresponding to keys in the JSON. It will
use \code{\link[jsonlite:read_json]{jsonlite::parse_json()}} and if it encounters an error, the error is discarded and
replaced with an empty list.
}
\description{
Note that you should really understand what is in that vector, because this function will discard
all errors and replace them with an empty list. The successfully converted vector will be a
vector of lists. Each list will have keys corresponding to the JSON keys. If you pass this a JSON
string with no keys (e.g. an array or number) it will probably work, but it wasn't built to do
that.
}
\keyword{internal}
