% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_responses.R
\name{split_responses}
\alias{split_responses}
\title{Split a responses table into semantically distinct parts.}
\usage{
split_responses(responses)
}
\arguments{
\item{responses}{A responses data frame (processed or not) from a CourseKata
class.}
}
\value{
A list of \code{\link{tibble}} objects with three elements:
\code{surveys}, for all responses from the course Pre- and Post-Surveys;
\code{quizzes}, for all responses from practice quizzes in the course; and
\code{in_text}, for all other items that appear in the textbook.
}
\description{
The resulting responses table can be quite large and is filled with many
different types of responses. Pass the responses table to \code{split_responses()}
to break the data into three distinct components: pre- and post-course
surveys, end of chapter quizzes, and in-text questions.
}
