% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/map_codebook.R
\docType{data}
\name{codebooks}
\alias{codebooks}
\alias{codebook_1.6}
\alias{map_codebook}
\title{Codebooks for surveys from Introductory Statistics: A Modeling Approach.}
\format{
Each codebook has the same structure, the reference ID for each
question in the codebook, the consistent variable name for that question,
and the prompt and response options (if any). These variables are as
follows:
\describe{
\item{lrn_question_reference}{the unique ID for the question on
Learnosity, and in the book; may or may not match the same question
across versions}
\item{var_name}{the unchanging name for the same question across
different versions of the book}
\item{prompt}{the prompt for the survey item}
\item{lrn_option_<0-n>}{where applicable, the available response options
for the question}
}
}
\usage{
codebook_1.6

map_codebook(object, quiet = FALSE, release = NULL)
}
\arguments{
\item{object}{The data frame to add variable names to.}

\item{quiet}{Should errors and messages be suppressed?}

\item{release}{A character vector of the codebook release versions  to use.
These are combined into a single look up table when the function is run.}
}
\value{
\code{map_codebook()} will convert the \code{object} to a \code{\link{tibble}}
and add a new variable, \code{var_name}, which gives a version independent name
to the to each item in the \code{object} with a matching
\code{lrn_question_reference} in the specified codebook(s).
}
\description{
There are multiple versions of Introductory Statistics: A Modeling Approach.
In each version there are some survey items that appear across different
versions of the book. These items are sometimes edited for typos, grouped
differently, or placed on different pages, though they largely remain
unchanged. When the items are edited on Learnosity, the item pool management
system used throughout the course, they are given a new reference ID, though
the question still asks the same question. To track the question across
different versions of the book, these codebooks provide consistent variable
names that do not change across versions. When you run any of the
\verb{process_*()} functions, the \code{map_responses()} function is called to add
these variable names to the data being processed.
}
\seealso{
\code{\link{process_responses}}, \code{\link{process_auxillary}}
}
\keyword{datasets}
