% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load file(s) from a directory, zipped directory, or file path.}
\usage{
load_data(object, regexp, class_id = NULL, limit_search = NULL)
}
\arguments{
\item{object}{The object to process the data from. This can be one of a
variety of options:
\itemize{
\item the path to a CourseKata data download zip file
\item the path to an extracted data download directory
\item the path to a specific CSV file (e.g. \code{classes.csv})
\item the \code{data.frame} loaded from a CSV file (useful for writing unit tests)
}}

\item{regexp}{If \code{object} is a directory or zipped directory, a
\code{\link{regex}} pattern indicating which file names to search for.}

\item{class_id}{A character vector of the class ID(s) to load.}

\item{limit_search}{If \code{object} is a directory or zipped directory, an
integer indicating how many files to match, e.g. \code{1} indicates that only
the first file found should be loaded. \code{NULL}, the default, uses all files
that \code{regexp} matches.}
}
\value{
If the \code{object} is a \code{data.frame} it is cast to a
\code{\link{tibble}}. If \code{object} is a path of some type, the files at the
given location are matched against \code{regexp}, \code{class_id}, and
\code{limit_search}, and then the matched files are are read in. As the files
are read, the rows of each table are appended to the rows of the previous
tables, yielding one merged table. The resulting data frame is returned as
a \code{tibble}.
}
\description{
Load file(s) from a directory, zipped directory, or file path.
}
