% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load file(s) from a directory, zipped directory, or file path.}
\usage{
load_data(path, regexp, class_id = NULL, limit_search = NULL)
}
\arguments{
\item{path}{The path to find the files at.}

\item{regexp}{If \code{path} is a directory or zipped directory, a
\code{\link{regex}} pattern indicating which file names to search for.}

\item{class_id}{A character vector of the class ID(s) to load.}

\item{limit_search}{If \code{path} is a directory or zipped directory, an
integer indicating how many files to match, e.g. \code{1} indicates that
only the first file found should be loaded. \code{NULL}, the default, uses
all files that \code{regexp} matches.}
}
\value{
A \code{\link{tibble}} of the data found at the path given, or if the
path given is a directory or zipped directory, the paths matching
\code{regexp}. If \code{limit_search} is specified, only that number of
files will be read and merged. The files are read using
\code{\link{map_dfr}} and \code{\link{read.csv}} with
\code{stringsAsFactors} set to \code{FALSE}.
}
\description{
Load file(s) from a directory, zipped directory, or file path.
}
