% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_auxillary.R
\name{process_auxillary}
\alias{process_auxillary}
\alias{process_classes}
\alias{process_page_views}
\alias{process_media_views}
\alias{process_items}
\alias{process_tags}
\title{Processing functions for all auxillary (non-response) data.}
\usage{
process_classes(object)

process_page_views(object, time_zone = "UTC", class_id = NULL)

process_media_views(object, time_zone = "UTC", class_id = NULL)

process_items(object, class_id = NULL)

process_tags(object, class_id = NULL)
}
\arguments{
\item{object}{The object to process the data from. This can be one of a
variety of options:
\itemize{
\item the path to a CourseKata data download zip file
\item the path to an extracted data download directory
\item the path to a specific CSV file (e.g. \code{classes.csv})
\item the \code{data.frame} loaded from a CSV file
}}

\item{time_zone}{The time zone to use when parsing date-time objects (see
\code{\link{timezones}}).}

\item{class_id}{(Optional) A character vector of the IDs of specific classes
you want to extract data for. These values should correspond to the names
of the directories where the desired CSV files can be found in the original
data download.}
}
\value{
These functions will always return a \code{\link{tibble}} of the
data you are processing, where each row describes the data named in the
function's name, e.g. \code{process_classes()} returns a \code{tibble} with rows
describing classes. For descriptions of each of variables in each of the
tables, see the README for this package (included in the data download or
online at \url{https://github.com/UCLATALL/CourseKataData}).
}
\description{
These functions are for individually processing the auxillary components of a
CourseKata data download. These functions assume that you are working files
downloaded from CourseKata, as these files have a predictable structure.
}
\details{
with a zip archive as downloaded from CourseKata, a directory extracted from
said zip archive, the path to the specif
}
\seealso{
Other processing functions: 
\code{\link{process_responses}()}
}
\concept{processing functions}
