% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process the data from a CourseKata data download.}
\usage{
process_data(path)
}
\arguments{
\item{path}{The path to a data download zip file or the path to an extracted
data download directory.}
}
\value{
This function returns \code{TRUE} if it completes successfully, and
\code{FALSE} if it is aborted.
}
\description{
This function loads and processes data from a CourseKata end-of-class
downloads. It functions by passing the path to each of the processing
functions in this package. The output data frames are all created in the
user's active environment as \code{classes}, \code{responses}, \code{items},
\code{page_views}, \code{media_views}, and \code{tags}. If any of the
variables already exist, the user is prompted to either allow overwriting or
abort the process (if not being run interactively, a message is emitted
noting that the variables were overwritten).
}
