% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_responses.R
\name{process_responses}
\alias{process_responses}
\title{Process responses from a CourseKata class data download}
\usage{
process_responses(object, time_zone = "UTC", class_id = NULL)
}
\arguments{
\item{object}{The object to process the data from. This can be one of a variety of options:
\itemize{
\item the path(s) to a CourseKata data download zip file or the directory extracted from that zip
\item the path(s) to a specific CSV file (e.g. \code{classes.csv})
\item a \code{\link{data.frame}}
}}

\item{time_zone}{The time zone to use when parsing date-time objects (see \code{\link{timezones}}).}

\item{class_id}{(Optional) A character vector of the IDs of specific classes you want to extract
data for. These values should correspond to the names of the directories where the desired CSV
files can be found in the original data download.}
}
\value{
A \code{\link{tibble}} of all responses found in the given path or data frame. If there are
multiple \code{responses.csv} files, they are merged together and can be distinguished by the
\code{class_id} column. If \code{object} is a directory and \code{class_id} is specified, only the responses
for the specified class or classes are included.
}
\description{
The \code{process_responses()} function processes the responses found in the given object. The
responses are checked for required columns in \code{ensure_data_in_responses()}, the variables in the
outcome table are converted to appropriate types with \code{convert_types_in_responses()}, and
multiple-response option indices in \code{response} (\code{lrn_type = 'mcq'}) are mapped to their values
(\verb{lrn_option_<index>}) with \code{map_response_options()}.
}
\seealso{
Other processing functions: 
\code{\link{process_auxiliary}}
}
\concept{processing functions}
